Added 'CyberSentinel' folder which includes modules for training a classification model.
Proposed a new set of modules specifically for training a binary classification model; includes a Dockerfile which pulls the latest TensorFlow image to containerize the training process. When it comes to this, I have no idea what I'm doing. So, expect the code to be crude, and far from testing.